+++ src/routes/user/components/scan_input.rs
@@ -3,10 +3,7 @@ use std::rc::Rc;
 use chrono::Utc;
 use leptos::{ev, leptos_dom::logging::console_log, prelude::*, task::spawn_local};
 
-use crate::{
-    models::UserId,
-    routes::{articles::get_article_by_barcode, user::MoneyArgs},
-};
+use crate::routes::{articles::get_article_by_barcode, user::MoneyArgs};
 
 use super::buy_article::buy_article;
 
@@ -14,7 +11,6 @@ pub fn invisible_scan_input(
     is_focused_signal: RwSignal<bool>,
     error_signal: RwSignal<String>,
     money_args: Rc<MoneyArgs>,
-    user_id: UserId,
 ) -> impl IntoView {
     let input_signal = RwSignal::new(String::new());
     let last_input = RwSignal::new(Utc::now());
@@ -77,8 +73,6 @@ pub fn invisible_scan_input(
                 return;
             }
 
-            // console_log(&format!("Seconds till last input: {} | Has passed 30s: {}", (Utc::now() - last_input.get()).num_seconds(), timediff()));
-
             // Clear input if nothing was typed for 30 seconds
             if timediff() {
                 input_signal.write_only().set(String::new());
@@ -93,13 +87,4 @@ pub fn invisible_scan_input(
     on_cleanup(move || {
         handle.remove();
     });
-
-    return view! {
-        // {
-        //     move || match is_focused_signal.get() {
-        //         true => console_log("input is focused"),
-        //         false => console_log("input is out of focus")
-        //     }
-        // }
-    };
 }
